Fix load/store flash content for ReproCartridgeV{1,2}
We accidentally constructed the AMDFlash object in the ReproCartridgeV1
and ReproCartridgeV2 devices with the property to don't load/save
modified flash content. This was probably caused by copy/pasting code
from the RomMatraInk device (where this property makes sense).

The problem was that this property is controlled via a boolean
constructor parameter, and that this unusual behavior isn't obvious from
looking at the caller. More in detail: the caller looks like this:
    flash(rom, sectorInfo, 0x207E, true, config, false)
And that last 'false' parameter indicates that the flash content should
not be loaded/saved.

This patch fixes the value of this last parameter for the
ReproCartridgeV1 and V2 devices. But it also changes the type of that
parameter from 'bool' to an i'enum'. That should make the meaning more
obvious and should hopefully prevent similar copy/paste problems in the
future.

Actually the 4th parameter is also a boolean, and from only looking at
the caller we also don't have any idea what it does. This patch changes
that to another enum. (It selects between '11-bit' or '12-bit'
'addressing', I refer to the code for more details on this).

The caller (in RomMatraInk) now looks like this:
    flash(rom, sectorInfo, 0x01A4, AmdFlash::Addressing::BITS_11,
          config, AmdFlash::Load::DONT)