Tweak VideoSystem::setClipboardText()
Pass a 'zstring_view' rather than 'string_view'. This gives the extra
guarantee that the string is zero-terminated. This avoids having to make
an extra copy of that string (in SDLVideoSystem).

On the (only) call-side we start from a TclObject (which guarantees
zero-terminated strings). So all we had to do is not loose the
information that the string is zero-terminated.