Fix two bugs in the RTC code
Thanks to Luca Chiodi (KdL) and HRA! for reporting and their BASIC test
program.

First problem was related to the leap-year counter. This counter is
stored in the 'alarm' block (instead of the 'time' block) of the RP5C01.
Also when changing this counter we must sync the registers with our
internal time representation. Before this patch we only synced when a
register in the 'time' block was changed.

The second bug was related to days-in-the-month. Suppose the current
date is '79/07/31/ and we want to change that to '79/12/31' (change the
month from July to December.

Because the RP5C01 works with nibbles this is a two step process. The
MSX-BASIC 'set date' command first writes the lower then the upper
nibble. So temporarily we set the invalid date '79/02/31' (February does
not have 31 days).

Before this patch our code would roll over that date from February to
March. But apparently the real RP5C01 only does days-in-the-month
adjustments when there was actually a carry from hour-roll-over into
days.