Fix compilation of tools in Contrib/ with gcc-11
We got an automated(?) email from the debian team that the openMSX
package would not build in their future release which will switch to
gcc-11.

The problem was not actually located in openMSX itself, but in the tools
we ship in Contrib/. The problem was a combination of:
 - gcc-11 defaults to c++17 mode.
 - c++17 has a new type 'std::byte'.
 - some of our code had both:
    * using namespace std'
    * typedef unsigned char byte;
And this caused an ambiguity between our 'byte' and 'std::byte'.

Fixed in two ways:
 - Use 'uint8_t' instead of 'byte'.
 - Don't use 'using namespace std', instead explicitly qualify all std
   types. (We already follow this rule in openMSX itself, but for the
   tools in Contrib we were more sloppy).
Either of these two by itself would have been sufficient. Doing both
makes the code more robust for possible similar problems in the future.