Implement reading from SCC for Carnivore2
This fixes SCC detection (e.g. in Nyyrikki's detection routine) when
using Carnivore2.

There was an explicit comment in our Carnivore2 emulation code:
   note: no reads from scc
I don't know if this was a TODO or if it was because older versions of
Carnivore2 really didn't support this.

In case case, I checked the Carnivore2 VHDL code, and today it does
support reads from SCC.