Fix report of 0kB memory mapper in MSXMEM.BAS on Sanyo_PHC-77
While working on adding a config for the 'Sanyo_PHC-77' machine we
noticed that MSXMEM.BAS reported a memory mapper of 0kB (on the real
machine). Before this patch there was no way to replicate that behavior
in openMSX.

At first it also wasn't clear where this strange behavior came from.
Because the machine really has no memory mapper. BiFi theorized that it
might be because the machine does have a S1985 which implements the
memory mapper IO ports (0xfc-0xff). We since verified that this
hypothesis is indeed correct.

This patch instantiates a dummy memory mapper when the S1985 device is
used and when the machine does not have a real memory mapper. The only
job of this dummy memory mapper is to ensure that the memory mapper IO
ports are available, but it doesn't show up in any MSX slot.

By itself this patch is not yet sufficient. We should still add a tag
like:
   <MapperReadBackBits>5</MapperReadBackBits> in the machine config(s).
I didn't do this yet because, according to the S1985 application manual,
the S1985 can operate in two modes: 4 or 5 bits memory mapper. I don't
know if this only controls how many memory mapper bits are put on the
output pins or that it also influences how many bits can be read back
via the memory mapper IO ports. IOW this needs more tests on real
machines.