S1985 overrules memory-mapper IO-port read behavior
This is a cleaner implementation of an earlier patch with title:
    Fix report of 0kB memory mapper in MSXMEM.BAS on Sanyo_PHC-77

Experiments have shown that on machines with a S1985, on reading the
memory-mapper IO ports:
- The lower 5 bits return the last written value. Even if the actual
  memory mapper is smaller than 512kB, or if there is no memory mapper
  at all.
- The upper 3 bits read as '100' on most machines, but we've also seen
  '000' on one specific machine (which one?). (*1)
- Adding an extra memory mapper in the system does not influence the
  previous two points. (Possibly the S1985 blocks the IO read request
  from appearing on the bus?)

(*1) This behavior can be selected via the <MapperReadBackBaseValue> tag
in S1985. But this patch does not set it in any machine config. And then
it defaults to 0x80, which sets the upper 3 bits as '100'.

After this patch the global <MapperReadBackBits> tag has no meaning
anymore. Instead the presence of the S1985 device itself fully specifies
the mapper-IO read-back behavior.
TODO this patch did not yet remove those tags from the machine configs.

Possibly in the future we can do something similar for the T9769B and
T9769C. (And possibly fully eliminate the <MapperReadBackBits> tag??)