Fix (theoretical) problem in debug build
In the previous patch we create a pointer one-past the last element in
an array. This is perfectly fine. However we do this via an expression
like '&*it', so dereference an stl-iterator and then take the address.

When compiling with a debug-stl-version this triggers a diagnostic
because dereferencing the end-iterator is not allowed (even though it
generates identical asm code as the current solution).

As a workaround we now do 'v.data() + (it - v.begin())'.

And all of this can be simplified when we switch to c++20, because it's
possible to construct a c++20 std::span directly from to iterators,
while our span can only be constructed from two pointers (hence we need
to convert iterators to pointers).

The simpler code is already present, but commented out.