Fix assert in planar-mode with VR=0
'Compile Disk-Station 5' occasionally triggered an assert. To reproduce
start the program and press space in the start-screen. You may need to
retry many times (+10) before you actually trigger it.

I found that it executes this code:
    ld a,(..)    ; a now contains 0x98, the VDP IO port
    ld c,a
    inc c
    xor a
    out (c),a
    ld (#fff6),a
    ld a,#97
    out (c),a
This is a good (standard) way to write VDP register 23. It allows for
non-default VDP ports and it updates the mirror system variables. But
the problem is that this code runs with interrupts enabled. In the
specific crash I was investigating there was an IRQ triggered between
the last two instructions in the above code.

The ISR also accesses VDP ports, this disturbs the above IO-sequence. So
when the ISR returns the following VDP accesses (including more accesses
not shown above) write bogus VDP registers.

As a consequence this selected screen mode G6 (screen 7) but with VR=0
(R#8 bit 3).

As far as I can tell, this is emulated correctly(?) in openMSX. You'll
see corrupted graphics, but I think this is how it should be? But it
also triggered an assert (only the assert was wrong, the actual code
seemed fine).