Store 'int' instead of 'string' in AfterCmd
Much of the openMSX OSD stuff (e.g. reverse-bar, os-menu, ...) is
implemented as Tcl scripts. Many of these are driven via repeated
'after' triggers. That is, we have an 'after' Tcl command that allows to
execute arbitrary Tcl code after some event happens (could be a
key-press, a video frame, some time-duration, ...).

This patch makes some tweaks in the implementation of 'after' so that
each 'after' command does one less heap-memory allocation. In some cases
we have a few after-commands per frame, and then this patch saves a few
hundred heap memory allocations per second.

More in detail. The Tcl interface for 'after' works more or less as
follows. You create an 'after-callback' (e.g. 'after frame do-stuff')
and this returns an 'id' for this new callback. With that 'id' it's then
possible to cancel the callback or query information on it.

This 'id' is a string of the form 'after#n' with 'n' some integer.
Before this patch this 'id' was stored as an actual string internally.
This patch instead stores an integer and only creates the string on
demand. Moreover when this (temporary) string is required it's created
on the stack rather than on the heap.