Tweak allocation in AfterCommand::executeMatches()
In the implementation of the 'after' command, when we receive an event,
we first filter all matching after-callbacks (we store those in a
temporary data-structure) and then execute all those matches. An
alternative could be to simply loop over all registered callbacks, check
if it matches and if so immediately execute it.

Unfortunately this (simpler) alternative doesn't work. The reason is
that executing a callback can execute arbitrary Tcl code. Including
registering extra or canceling existing callbacks. And changing the
vector of registered callbacks while looping over it causes all kinds of
technical problems. Hence the extra step of first filtering into a
temporary data-structure and only then executing the matches.

This patch tweaks how that temporary data-structure is allocated. Before
this patch, for each event-delivery, it was allocated on the heap and
then freed. With this patch we still allocate it on the heap, but we
don't free it anymore. Instead we keep the allocated buffer around for
the next event-delivery. We only free and reallocate a new buffer when
the previous one wasn't big enough (so very quickly we effectively won't
allocate anymore).

In some cases this patch saves hundreds of heap-allocations per second.