Use std::variant in AfterCommand
The previous 2 patches already reduced heap-allocations for the 'after'
command. This patch finishes that work. Now (after a short while, when
we reach steady-state) there are no longer any heap-allocations required
for 'after' command handling.

There are various sub-types of 'after' callbacks. We can request a
callback after some amount of emulation-time has passed, after some
real-time, after a specific input-event, after a frame has been
rendered, ... Before this patch we used a class-hierarchy to represent
these sub-types. And thus a specific after-callback maps to an object
in this class-hierarchy.

This is a classical OO (object-oriented) solution with lots of small
objects. In C++ these objects typically all need to be allocated on the
heap.

An alternative solutions is to replace the class-hierarchy with a
std::variant. std::variant is new in c++17. In other programming
languages a variant might be knows as a discriminated union, a tagged
union, a sum-type, or an enum-type (in rust). In short: a std::variant
is like a union but with some extra information to indicate which field
in the union is 'active'. Virtual methods in the class hierarchy are
replaced with std::visit() function call. For more details see:
https://en.cppreference.com/w/cpp/utility/variant

In this case our variant stores either:
 a AfterTimeCmd object, OR
 a AfterIdleCmd object, OR
 a AfterSimpleEventCmd object, OR
 a AfterInputEventCmd object, OR
 a AfterRealTimeCmd
PLUS some indication for which of these 5 types it actually holds (so
typically one extra byte).

That means that our variant is as large as the largest of these 5 types,
plus one byte (plus possibly some padding). So a variant works best if
the size-difference between the largest and the smallest type is not too
big (that's fine in this case).

Because all after-command objects are now equal in size, we can store
them in some instance of our ObjectPool class. (Technically it's also
possible to store variable-sized objects in a pool, but that requires a
more complex pool implementation, approaching the complexity of the
generic heap-allocator).

And then because we store these object in an ObjectPool rather than on
the heap, as soon as this pool has grown to a large enough size (in
practice the initial size (=256 elements) is already large enough) we
don't perform any heap allocations anymore.