Replace Event::operator< with operator==
All Event(sub)classes implement operator< and  operator== is implemented
in terms of operator<:
  operator==(a, b) : !(a < b) && !(b < a)
So operator== (possibly) calls operator< twice.

Probably historically operator< was required (maybe because we once
stored events in a std::map?). But in the current code only operator==
is (still) needed.

This patch replaces Event::operator< with Event::operator==. This is
faster because:
 * we no longer call operator< twice
 * typically an operator== implementation can be simpler than an
   operator< implementation.