Use std::variant and ObjectPool for Event
Some recent commits already reduced the amount of recurring heap memory
allocations. Another source of such allocations are events, for example
input events. In case the mouse (and/or joystick) is moved, this
generates dozens or even hundreds of events per second. Before this
patch each such event caused an instantiation of a small polymorphic
object from some class hierarchy. All such objects had to be allocated
on the heap.

This patch changes the Event class-hierarchy to a std::variant. And
allocates those from an ObjectPool. Similar to how the AfterCmd
hierarchy was changed in an earlier patch. This eliminates practically
all Event-related heap-allocations. (I measured that for typical use
cases the pool-size never grows beyond 100 simultaneously live objects.
In the current implementation that only takes one memory chunk).

Before this patch the 'Event' class was the base-class for all events.
All events were passed/stored as 'shared_ptr<const Event>'. After this
patch we still have an 'Event' class, but now it's a handle-class. It
wraps a pointer to some (hidden) implementation class. This internal
class contains a reference-count plus a std::variant that stores the
actual (sub-)event information.

In some cases we pass Events from the non-main- to the main-thread. So
this patch contains some amount of locking to enable this (before this
locking was located inside the implementation of shared_ptr and the
global new/delete).

std::variant works best when the size-difference between the largest and
the smallest type in the variant is not too large. The next few patches
reduce that maximum size to 24 bytes (now it's still 40 bytes), the
smallest size is 8 bytes. Together with the ID stored in std::variant
itself and the reference count we added, this brings the size of (all)
event-objects to only 32 bytes. Before this patch even the smallest Event
object already took 32 bytes (the vtable pointer, the EventType enum,
the shared_ptr reference count, the heap allocation overhead, ...). So
even though we now use a variant (and all events take up the same space
as the largest event), we still use strictly less memory than before.

This is a relatively large patch, though most of it are boring
mechanical changes. Most interesting parts are located in
'src/events/Event.{hh,cc}'.

Note: an alternative which I considered was to store/pass Event objects
by value instead of allocating them (from a pool or from the heap). This
would be more efficient. But it's not compatible with
reference-counting, that's not a big problem, we could do without
ref-counts. But a real problem is that this is not compatible with
self-references: an OsdControlEvent has a reference to another Event as
one of its members.