Reduce size of GroupEvent by using static_vector
This matters because GroupEvent was the largest type in 'EventVariant'.
So reducing the size of GroupEvent reduces the size of all Events.

GroupEvent contains a 'std::vector<EventType>' that holds the types of
all the allowed events in a group. In the current code we never need
more than 2 different types. This patch changes the type from
std::vector to 'static_vector<EventType, 3>'. (So we allow up to 3 types,
because of padding this doesn't cost more than 2. Increasing to 11
should also be possible without increasing the size of EventVariant).

Also tweaked static_vector to store the current size in the smallest
possible integer type. E.g. if the vector can hold 200 elements, then an
8-bit integer is sufficient.

With this tweak a 'static_vector<EventType, 3>' only takes 4 bytes:
* 3 bytes for the buffer
* 1 byte to indicate how many positions in the buffer are occupied.

In contrast: a std::vector object already takes 24 bytes (3 pointers),
even if the vector is empty. A non-empty std::vector additionally
allocates a buffer on the heap (which typically takes a minimum of 16
bytes).