Tweak 'openmsx_info setting' for proxy settings
A proxy setting (e.g. 'firmwareswitch') refers to the machine-specific
setting of the currently active machine instance. It's possible that one
instance has and another instance does not have a particular machine
setting.

The 'openmsx_info setting' command behaved inconsistent in this case. On
a query like
   openmsx_info setting firmwareswitch
with:
* no machine instance has this setting: it returns an error
* the active machine has this setting: it returns the expected values
* the active machine does not have this setting: it returned dummy
  values.

This patch changes the behavior of this last case. This will now also
return an error. In other words, the behavior of proxy-settings should
only be influenced by the active machine instance, not by random other
instances.

This is a followup on commit d762d1d064cf. That one ran into the above
issue, and worked around it. That workaround is still valid (maybe even
a more direct way), so I'm not going to revert that patch.