Fix osd menu with machine tabs with different number of disk drives
To reproduce:
- Start openMSX with a machine with 2 disk drives.
- Open a 2nd tab with a machine with only 1 drive:
   menu > Advanced > Manage Running Machine > New Running Machine Tab
- Close and reopen the menu.
This triggers an error because it tries to execute the 'diskb' command
in the machine instance that does not have a B-drive.

Internally the menu-code executes the command
    info command disk?
to query which drives (and similar for cartridge slots) are available.
Notice it's using the native Tcl 'info' command (not the 'openmsx_info'
variant). So we can't teach it to only return commands for the active
machine instance.

Fixed by filtering this list: only keep those commands that actually
work in the current machine instance.