[1/3] remove speedSetting from Autofire::getSignal()
Implement Autofire::getSignal() solely based on 'clock'. Before it also
checked the value of 'speedSetting'.

More in detail: 'speedSetting' sets the frequency of 'clock'. So in
principle 'clock' alone should contain enough information. Except for
the special value '0', this disables Autofire. Before this patch
'getSignal()' would first check 'speedSetting' for '0' and only then
query 'clock'. Now this special value is encoded in 'clock' (by setting
clock-period = 0).

This change is a preparation to fix autofire in combination with
'reverse-replay'. During a replay the active autofire-rate can be
different from the current value of the setting.