Extended CasImage to also support SVI .cas files
Before it only understood MSX .cas files.

This patch both simplifies and generalizes the CasImage class:
- Generalize: before the frequency of the generated wave data was fixed
  (set at compile-time). Now it's dynamic (set at run-time).
- Simplify: move all of the .cas file parsing code out of the class. All
  that remains is a buffer of the final wave data and the frequency of
  this data.

The CasImage constructor checks the first few bytes of the .cas file and
then parses it either as a SVI or as a MSX cas file. (And this parsing
code is implemented outside of the CasImage classr. Bonus: the result
can now be made 'const').