READ-DMK: enlarge 'sector_buffer'
It was already bigger than the total length of the track. But
nevertheless when there are lots of overlapping sectors in a
(copy-protected) track, then this buffer could overflow (luckily this
causes the tool to stop with a clean error, rather than crash).

This patch enlarges the buffer with 4 kB.

See also this MRC thread:
    https://msx.org/forum/msx-talk/software/supercop-erbosoft