Generalize TC8566AF emulation
The next patch adds the 'Talent TDC-600' external floppy disk interface.
That interface uses the WD37C65 disk controller. The WD37C65 and
TC8566AF are mostly compatible (at least for all things relevant for
MSX, maybe not for stuff like DMA and high-density disk support).
Actually both seem to be clones of the NEC765 FDC and are compatible
with that one (instead of directly compatible with each other).

The previous paragraph states that the WD37C65 and TC8566AF are (mostly)
compatible. You program the control-, data- and status-register in the
same way. However the way to access those registers does differ. This
patch generalizes the current TC8566AF emulation code to make
abstraction from how exactly to access those registers.

For example, before this patch, to read the status register, you did
something like:
    tc8566af.readReg(4, time);
This patch changes that to:
    tc8566af.readStatus(time);

For now I'm keeping the name 'TC8566AF' as the name of this class. But
actually it's now more generic than only TC8566AF emulation.