Add Talent_TDC-600 emulation (doesn't work yet)
This patch adds emulation of the TDC-600 external floppy disk interface.
The comments in the code contain a link to the electric schema, from
that it was possible to deduce the exact ROM and register mapping (and
mirroring).

However the interface doesn't actually work yet. That's because at the
moment the TC8566AF emulation isn't accurate enough yet: it's sufficient
for e.g. the software in the turboR disk ROM, but not for the TDC-600
software. (Actually TC8566AF emulation was tweaked until the turboR
started working, details that weren't important for that use case
weren't considered).

The next patch will improve TC8566AF emulation so that both turboR and
TDC-600 software work.