Improve emulation of TC8566AF seek/recalibrate commands
After this the TDC-600 disk rom software works fine.

The main changes are in the seek and the recalibrate (=seek to track 0)
commands. Before these commands were modelled similar to the other
commands (and similar to the corresponding WD2793 commands). That is:
the command finishes when the seek operation finishes. Instead on a real
TC8566AF these commands finish immediately and the actual seek continues
in the background. The advantage of this is that it allows to execute
simultaneous seek commands on all connected disk drives. Or even e.g.
one read-command simultaneous with a seek on another drive.

None of the MSX disk rom software uses this simultaneous-seek feature.
Nevertheless it was needed to (better) emulate this to make the TDC-600
software work. The reason for this is a difference in how e.g. the
turboR disk rom and the TDC-600 disk rom check for the end of a seek:

The turboR disk ROM software will:
a) issue the seek command
b) wait for the end of this command
c1) in a loop execute the 'sense interrupt status command'
c2) from the result of this command poll the 'Seek-End' status bit

The TDC-600 software did the same, but without step 'b'. On real
hardware that's fine because the seek command 'finishes' immediately.

Before this patch, the 'turboR' version happened to work fine (but the
wait happens in step b rather than in step c). But on TDC-600 (with the
wrong seek emulation code) the 'sense interrupt status' command got
issued before the seek command was finished. And this caused the
poll-loop in step 'c' to check for a bit that would never change. So
effectively hang.