Change parameter-type passed to (virtual) help() methods
From:
    help(const std::vector<std::string>& tokens);
to:
    help(span<const TclObject> tokens);

This simplifies/optimizes the (single) caller of this method from:
    auto tokens2 = to_vector(view::transform(
        view::drop(tokens, 1),
        [](auto& t) { return string(t.getString()); }));
    help(tokens2);
to:
    help(tokens.subspan(1));

So no need to convert TclObject to std::string, and no need to allocate
a vector to hold these temporary strings. Of course help() is never a
performance critical operation. But it's a nice cleanup anyway (both on
the caller and callee side).

The old parameter type (vector<string>) was still there for historic
reason (probably still comes from the time before we used TclObject).
There was never any (performance) reason to change this, so I never
bothered. But with some global search/replace still wasn't so much work
after all.

There is a corresponding (virtual) execute() method. That was changed to
'span<const TclObject>' already a long time ago (because that does
matter for performance). So this patch also makes help() and execute()
more similar.