Fix compilation error on clang (related to CTAD and initializer_list)
Worked fine on gcc and msvc, not on clang.

See patch for details. I would have expected this to work (starting from
c++17). But then I also couldn't find anything on this in the
documentation (in a few minutes). So maybe clang is right?

Anyway, easy enough to fix/work around.