Change find_if() to find() by using projection functions
c++20 added 'projection functions' to many algorithms. We cannot use
c++20 yet in openMSX, but we can 'import' useful ideas into our code
base.

So what is a 'projection function'? Many STL algorithms operate on a
range of values (e.g. on a vector-of-ints or on an array-of-structs).
Many of these algorithms now take an extra (optional) parameter which is
a 'projection function'. This function 'transforms' these values before
the algorithm operates on them.

It's easier to explain via an example:
    struct Person {
        std::string firstName;
        std::string lastName;
        int age;
        ...
    };
    std::vector<Person> persons;
    ...
    auto it = std::ranges::find(persons, "Joe", [](const Person& p) { return p.firstName; });

Without projection function (without 3rd parameter), the 'find()'
algorithm would operate on 'Person' objects, and thus search for a
Person that is equal to some other Person (assuming there is an
'operator==(Person, Person)'). However in this example we only want to
check the 'firstName' field. So we pass a function that takes a 'Person'
object as input and only selects the 'firstName' field from this. This
behavior is very typical: take a whole struct and return only part of
it. Mathematically this kind of operation is called a 'projection', so
the name 'projection function' is well chosen.

Very often the projection function selects a single field. For this
common case there is a shorthand notation:
    auto it = std::ranges::find(persons, "Joe", &Person::firstName);
Technically '&Person::firstName' is a pointer-to-member-variable,
and the logic to interpret this as a function that returns this field is
implemented in the std::invoke() function. For lots more details see:
    https://en.cppreference.com/w/cpp/utility/functional/invoke

Note that projection functions don't enable anything new. It was already
possible to get the same effect with the 'find_if()' algorithm:
 * find(): search for something that's equal-to a given value
 * find_if(): search for something that matches a given predicate
The example then becomes:
    auto it = std::ranges::find_if(persons, [](const Person& p) { return p.firstName == "Joe"; });
So 'find_if()' is strictly more powerful than 'find()', though because
of that:
 * Using find() is often shorter to write, especially when using the
   shorthand-notation.
 * Because find() is less generic, it conveys more semantic meaning:
   - With find() you immediately know we're looking for an equality.
   - With find_if() you have to read the predicate to figure this out.

So far the example only focused on the 'find()' algorithm (because
that's easiest to explain). But in algorithms like 'sort()' (and binary
search algorithms) using projection functions is often even more
convenient. Compare:
    std::ranges::sort(persons, [](const Person& p1, const Person& p2) { return p1.age > p2.age; });
with:
    std::ranges::sort(persons, std::greater{}, &Person::age);
Note:
 * The selection of the 'age' field is repeated twice in the first
   version.
 * The first version mixes the order (big-to-small) with the
   field-selection. In the second version these two are separated.
 * Most often you want to sort from small-to-big, in that case you can
   pass '{}' as the second parameter (defaults to 'std::less{}').

Ok, that was a long explanation of 'project functions'. Now about this
patch itself:

This patch adds an optional projection-function parameter to our
'find()' algorithm and then uses it where possible. Often this turns a
former 'find_if()' into a 'find()'.

Later patches is this series also add projection-functions to other
algorithms.