Add (optional) projection-function to sort() algorithm
For now this patch only uses this new function in one place. Later
patches in this series will add additional callers (but this first needs
some other cleanups).

In c++17 it's not yet (easily) possible to make 'constexpr' projection
functions (technically because std::invoke() is not yet constexpr in
c++17), in c++20 this is all fixed. Because of this there are a few
places where we couldn't yet use projection in sort() where it would
actually be useful.