Add projection-function parameter to find_unguarded() and rfind_unguarded()
Similar to how an earlier patch added such a parameter to 'find()' and
then replaced 'find_if()' calls.

In a few places this patch makes some changes that only look marginally
simpler. For example:

Before:
    rfind_if_unguarded(sources, [&](auto& p) { return p.second == id; });
After:
    rfind_unguarded(sources, id, [](auto& p) { return p.second; });

Though this is also a preparation to in a later patch simplify this
further to:
    rfind_unguarded(sources, id, &Source::id);