Add (optional) projection-function to is_sorted()
'is_sorted()' is not used that often in our code base. But one of the uses
is a good example for how projection functions can simplify the code.

Before:
    struct AscendingZ {
        bool operator()(const unique_ptr<OSDWidget>& lhs,
                        const unique_ptr<OSDWidget>& rhs) const {
            return lhs->getZ() < rhs->getZ();
        }
    };
    assert(ranges::is_sorted(subWidgets, AscendingZ()));
After:
    assert(ranges::is_sorted(subWidgets, {}, &OSDWidget::getZ));
7 lines replaced by a single line (though in the actual code the
'AscendingZ' functor is used in 2 locations).

The first patch in this series already mentioned the shorthand notation
for project functions. This example uses a different variation. Notice
how 'OSDWidget::getZ()' is a method (not a variable). So it's not only
possible to use shorthand notation to select fields from a struct but
also to call nullary-member-functions.

Also notice how in this example the 'getZ()' function is called via
'unique_ptr'. The earlier examples selected fields from pass
by-reference objects, here the objects are passed via (smart-)pointer.
All these variations are handled by std::invoke(). So the same
shorthand-syntax '&Foo::bar' could mean either:
 * select a field in a value-object:  object.bar
 * select a field via a pointer:      object->bar   (regular- or smart-pointer)
 * call a method on a value-object:   object.bar()
 * call a method via a pointer:       object->bar() (regular- or smart-pointer)