Add projection-function to lower_bound(), upper_bound() and equal_range()
These three algorithms perform a binary-search in a sorted range
(possibly containing duplicates):
* lower_bound: returns the earliest position where the given value could
               be inserted while keeping the range sorted.
* upper_bound: return the last position where the given value could be
               inserted while keeping the range sorted.
* equal_range: return both of the previous 2 results (but more efficient
               than calling these two in sequence).

In openMSX we often use these for fast (log(n)) lookup in a sorted
vector. For example:
    struct Foo {
        int key;
        ... // many other fields
    };
    vector<Foo> v; // sorted on Foo::key
    ...
    auto it = lower_bound(v, myKey, {}, &Foo::key);
    if (it != end(v) && it->key == myKey) { ...

When using projection functions this is fairly straight-forward: search
in 'v' for 'myKey', but only compare the 'key' field in 'Foo'. (And the
resulting iterator still points to the full 'Foo' object, not just the
key part in it).

Before this patch this was already possible by using a custom comparison
function (the example above uses the default 'std::less{}' ordering, 3rd
parameter). The lower_bound() call then looked like this:
    auto it = lower_bound(v, myKey, [](const Foo& x, int y) { return x.key < y; });
This was certainly more complex. Also notice how this is a heterogeneous
comparison: the left-hand-side is an element from the range and the
right-hand-side is key-value we want to locate.

For upper_bound() the types of the left- and right-hand-sides are
swapped. And equal_range() even needs both orders! (Not possible with a
simple lambda, needs a functor with overloaded operator(), e.g. see the,
now removed, 'CompareBreakpoints' functor in this patch). There are good
technical reasons for these different orderings, but as a user of these
algorithms I prefer to not have to know these details. And by using
projection functions I no longer have to know.

This patch extends the 3 binary-search algorithms and then uses them in
~10 locations. Later patches in this series will change more uses (but
these first require other cleanups).