Simplify OptionData in CommandLineParser
Before this patch we had:
    struct OptionData {
        CLIOption* option;
        ... // various other fields
    };
    vector<pair<string_view, OptionData>> options; // sorted on pair.first

This data-structure contains information on command line options.
OptionData contains various information on how to handle the option. The
name of the option is stored as the first field in a pair, OptionData is
the second field of the pair. All these pairs are stored in the 'options'
vector which is sorted on pair.first (= the name).

This patch simplifies that to:
    struct OptionData {
        string_view name;
        CLIOption* option;
        ... // various other fields
    };
    vector<OptionData> options; // sorted on OptionData::name
So the name is moved from the pair into OptionData, and then the pair is
no longer needed, so the vector directly contains OptionData.

This has two main advantages:
* Instead of using 'first' and 'second' to access the fields in the
  pair, we can use more meaningful names. For example:
   - it->first          becomes  it->name
   - it->second.option  becomes  it->option
* We can now use the shorthand-notation '&OptionData::name' in
  lower_bound() and sort() algorithms. (Technically also possible for
  '&pair<string_view, OptionData>::first' but that's no longer a
  _shorthand_ notation ;-)

Actually the reason why we initially used a pair was because we had a
helper functor 'LessTupleElement' which made the syntax for
lower_bound() searches with pair (or tuples) still reasonably clean. I
did not know how to write this helper for structs. But this new solution
is still cleaner.

The next patches in this series replace more pairs with structs with
properly named fields.