Tweak IPSPatch
Similar to previous patch, replace:
    std::pair<size_t, std::vector<byte>>
with:
    struct Chunk {
        size_t startAddress;
        std::vector<byte> content;
    };
This is exactly the same information, but the latter allows to give more
meaningful names to stuff (e.g. 'it->first' vs 'it->startAddress').

This also allows to use cleaner 'projection' syntax in the
range-algorithms. For example compare:
    std::ranges::lower_bound(chunks, src, {}, [](const auto& p) { return p.first; });
with:
    std::ranges::lower_bound(chunks, src, {}, &Chunk::startAddress);

This commit also turns IPSPatch into an immutable class. That is, after
construction all member variables remain unchanged. This already was the
case before this commit, but now it's obvious because all member
variables (and member functions) are 'const'.