Fix 'after frame' after ea1da72bac15c
The 'after frame' and 'after quit' commands didn't work anymore after
commit ea1da72bac15c. This patch fixes that.

The reason was that, even though these events are handled by the
'AfterCommand::executeSimpleEvents()' method (which is fine), the
'FinishFrameEvent' and 'QuitEvent' classes don't inherit from the
'SimpleEvent' class. And because of this the std::visit() dispatch
mechanism didn't work anymore because it only delegated actual
'SimpleEvent's to the 'executeSimpleEvents()' method. Fixed by
explicitly enumerating 'FinishFrameEvent' and 'QuitEvent'.