[1/10] string->id in AutofireState and KeyJoyState
When the user presses/releases a key on the key-joystick, this gets
recorded (for later replay) in a KeyJoyState object. We support 2
key-joystick input devices, so a KeyJoyState object needs to store on
which of the two key-joystick devices it applies.

Before this patch this information was stored as a string, the name of
the key-joystick. This patch changes that to an enum.

This is a trade-off. A string is more general: we don't need to
enumerate beforehand which key-joystick devices will ever be created. On
the other hand a string takes a lot more space and comparing a string is
more expensive than comparing an enum.

For later patches in this series the size of StateChange objects will be
important. So this patch is a preparation for that.

There's a similar change for Autofire. AutofireState objects also
recorded (via a string) on which Autofire they apply. But in the current
code there is at most one Autofire object: renshaturbo. This patch keeps
the (currently redundant) information, but changes it from a string to
an enum.

Before this patch sizeof(KeyJoyState) and sizeof(AutoFireStateChange)
was both 56, after this patch the size is reduced to only 24.